Divide subtitleTrackBody from defaultTrackBody
